# src/entity/enums.ts
export enum Status {
    INIT = "INIT",
    SUCCESS = "SUCCESS",
    FAILED = "FAILED",
}

export enum Unite {
    BOUT_PLAST = "BOUTEILLE PLASTIQUE",
    BOUT_VERRE = "BOUTEILLE EN VERRE",
    CASIER = "CASIER",
    PALETTE = "PALETTE",
    BRIQUE = "BRIQUE",
    CARTON = "CARTON"
}

export enum TypeVente {
    VENTE = "VENTE_SIMPLE",
    COMMANDE = "COMMANDE"
}


export enum MoyenPaiement {
    CASH = "CASH",
    PAIE_CASH = "PAIE_CASH",
    OM = "ORANGE MONEY",
    MOMO = "MTN MOBILE MONEY"
}